search: Add "x" button to the filter of uploaded files table.
Adds an "x" button to the filter input for uploaded files to allow users
to clear the filter easily.

A new function is introduced in `list-widget.ts` to handle user actions
when clicking the `clear_filter` button. This function will be utilized
in future commits.
Additionally, the CSS for the input filter has been updated.